// src/models/avatar-configuration-options-model.ts
export interface AvatarConfigurationOptionsModel {
    hairConfigurations: { id: string, name: string }[];
    hairColors: { id: string, name: string }[];
    skinColors: { id: string, name: string }[];
    facialHairConfigurations: { id: string, name: string }[];
    clothings: { id: string, name: string }[];
}